make output optional

+++ modules/nf-core/rundbcan/easycgc/main.nf
@@ -20,7 +20,7 @@ process RUNDBCAN_EASYCGC {
     tuple val(meta), path("${prefix}_cgc.gff")                 , emit: cgc_gff
     tuple val(meta), path("${prefix}_cgc_standard_out.tsv")    , emit: cgc_standard_out
     tuple val(meta), path("${prefix}_diamond.out.tc")          , emit: diamond_out_tc
-    tuple val(meta), path("${prefix}_TF_hmm_results.tsv")      , emit: tf_hmm_results
+    tuple val(meta), path("${prefix}_TF_hmm_results.tsv")      , emit: tf_hmm_results, optional: true
     tuple val(meta), path("${prefix}_STP_hmm_results.tsv")     , emit: stp_hmm_results
     path  "versions.yml"                                       , emit: versions
 
@@ -49,9 +49,11 @@ process RUNDBCAN_EASYCGC {
     mv cgc.gff                  ${prefix}_cgc.gff
     mv cgc_standard_out.tsv     ${prefix}_cgc_standard_out.tsv
     mv diamond.out.tc           ${prefix}_diamond.out.tc
-    mv TF_hmm_results.tsv       ${prefix}_TF_hmm_results.tsv
     mv STP_hmm_results.tsv      ${prefix}_STP_hmm_results.tsv
     mv total_cgc_info.tsv       ${prefix}_total_cgc_info.tsv
+    if [ -f TF_hmm_results.tsv ]; then
+        mv TF_hmm_results.tsv   ${prefix}_TF_hmm_results.tsv
+    fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
